refact: move cluster consistent access to be part in the defult manager
